Move acpi functions into module
Will be splitting this out into linux acpi functions
and windows acpi functions and just getting the clutter
out of the main file.
